requirements.txt
asgiref==3.5.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
crispy-bootstrap5==0.6
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-crispy-forms==1.14.0
django-extensions==3.2.0
django-filter==21.1
django-heroku==0.3.1
django-nocaptcha-recaptcha==0.0.20
django-recaptcha==3.0.0
django-widget-tweaks==1.4.2
extensions==0.4
filelock==3.7.1
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
oauth2client==4.1.3
oauthlib==3.2.0
Pillow==9.1.0
platformdirs==2.5.2
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-decouple==3.6
pytz==2018.5
readchar==3.0.5
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
selenium==3.14.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
whitenoise==6.0.0
